Improve test coverage
There was a hidden source of bugs because the user being created in the valid signup test had password: 'password1', while the sign_in_as method injected 'password'. Therefore logging-in failed as expected, but because the password was incorrect, not because the user hadn't activated their account.
The test for deletion of non-existant posts also produced false positive because the post was failing at the logged_in_user filter, before reaching the correct_user filter. Logging in first was the fix.
Deleted 2 files related to application_cable, which isn't used, and one related to application_job, which isn't either.
Finally, added imperfect test which goes some way to ensuring the cost of password hashing in production is sufficiently high. But I was unable to fool ActiveModel::SecurePassword into setting #min_cost itself, so the test sets this manually.