Redirect to homepage, not user page after login
This follows a convention established in other sites. Also, we can expect most users to want to first catch up on posts in their feed most times they login.
Had to update an integration test that was to specific about the redirect used, and the template rendered. Probably best not to assert_template; it tests implementation rather than behaviour.