Update README and LICENSE
I wrote the majority of the codebase, (as opposed to copying the code available in http://railstutorial.org/). So it seemed appropriate that Michael Hartl's licences and attributions should apply only to the copied code (much of which is boilerplate), while the remainder will be under the MIT licence alone.